+++ t/apicast-policy-chains.t
@@ -177,3 +177,61 @@ yay, api backend
 --- error_code: 200
 --- no_error_log
 [error]
+
+
+
+=== TEST 4: policy chain with invalid elements
+Invalid policies are removed from the chain.
+--- http_config
+  include $TEST_NGINX_UPSTREAM_CONFIG;
+  lua_package_path "$TEST_NGINX_LUA_PATH";
+  init_by_lua_block {
+    require('apicast.configuration_loader').mock({
+      services = {
+        {
+          id = 42,
+          backend_version = 1,
+          backend_authentication_type = 'service_token',
+          backend_authentication_value = 'token-value',
+          proxy = {
+            policy_chain = {
+              { name = 'apicast.policy.phase_logger' },
+              { name = 'invalid' },
+              { name = 'apicast.policy.echo' }
+            },
+            api_backend = "http://127.0.0.1:$TEST_NGINX_SERVER_PORT/api-backend/",
+            proxy_rules = {
+              {
+                http_method = "GET",
+                pattern = "/test",
+                metric_system_name = "hits",
+                delta = 1
+              }
+            }
+          }
+        }
+      }
+    })
+  }
+
+--- config
+  include $TEST_NGINX_APICAST_CONFIG;
+
+  location /transactions/authrep.xml {
+    content_by_lua_block { ngx.exit(200) }
+  }
+--- request
+GET /test
+--- response_body
+GET /test HTTP/1.1
+--- error_log chomp
+running phase: rewrite
+running phase: access
+running phase: content
+running phase: header_filter
+running phase: body_filter
+running phase: post_action
+running phase: log
+--- error_code: 200
+--- no_error_log
+[error]